+++ src/components/search-modal/search-badge.js
@@ -1,30 +1,27 @@
+import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
 
-import {
-  IconButton
-} from '@folio/stripes/components';
-import { FormattedMessage } from 'react-intl';
+import { IconButton } from '@folio/stripes/components';
 
-export default function SearchBadge({ filterCount = 0, onClick, ...rest }) {
-  return (
-    <FormattedMessage id="ui-eholdings.filter.togglePane">
-      {([ariaLabel]) => (
-        <div {...rest}>
-          <IconButton
-            icon="search"
-            onClick={onClick}
-            data-test-eholdings-search-filters="icon"
-            data-testid="search-badge"
-            {...filterCount > 0 ? { badgeCount: filterCount } : {}}
-            ariaLabel={ariaLabel}
-          />
-        </div>
-      )}
-    </FormattedMessage>
-  );
-}
+const SearchBadge = ({ filterCount = 0, onClick, ...rest }) => (
+  <FormattedMessage id="ui-eholdings.filter.togglePane">
+    {([ariaLabel]) => (
+      <div {...rest}>
+        <IconButton
+          icon="search"
+          onClick={onClick}
+          data-testid="search-badge"
+          {...(filterCount > 0 ? { badgeCount: filterCount } : {})}
+          ariaLabel={ariaLabel}
+        />
+      </div>
+    )}
+  </FormattedMessage>
+);
 
 SearchBadge.propTypes = {
   filterCount: PropTypes.number,
   onClick: PropTypes.func.isRequired,
 };
+
+export default SearchBadge;