+++ src/routes/package-show-route/package-show-route.js
@@ -15,10 +15,14 @@ import {
   FIRST_PAGE,
   INTERVAL_BEFORE_CHECK_FOR_AN_UPDATE,
   tagPaths,
+  titleSortFilterConfig,
+  selectionStatusFilterConfig,
+  publicationTypeTitlesListFilterConfig,
 } from '../../constants';
 
 import View from '../../components/package/show';
 import { SearchSection } from '../../components/search-section';
+import TitleSearchFilters from '../../components/title-search-filters';
 
 class PackageShowRoute extends Component {
   static propTypes = {
@@ -370,6 +374,19 @@ class PackageShowRoute extends Component {
     getPackageTitles({ packageId, params });
   };
 
+  renderSearchSectionFilters = (props) => {
+    return (
+      <TitleSearchFilters
+        availableFilters={[
+          titleSortFilterConfig,
+          selectionStatusFilterConfig,
+          publicationTypeTitlesListFilterConfig,
+        ]}
+        {...props}
+      />
+    );
+  };
+
   render() {
     const {
       history,
@@ -435,6 +452,7 @@ class PackageShowRoute extends Component {
               searchType={listTypes.TITLES}
               onFilter={this.searchTitles}
               onToggleActions={this.handleAccordionHeaderSearchActionsToggle}
+              renderFilters={this.renderSearchSectionFilters}
               {...props}
             />
           )}